requirements/prod.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements/prod.in
#
certifi==2022.9.24
    # via requests
charset-normalizer==2.1.1
    # via requests
confluent-kafka==1.9.2
    # via -r requirements/prod.in
h5py==3.7.0
    # via -r requirements/prod.in
idna==3.4
    # via requests
numpy==1.23.4
    # via
    #   -r requirements/prod.in
    #   h5py
    #   torchvision
pillow==9.3.0
    # via torchvision
requests==2.28.1
    # via torchvision
torch==1.13.0
    # via
    #   -r requirements/prod.in
    #   torchvision
torchvision==0.14.0
    # via -r requirements/prod.in
typing-extensions==4.4.0
    # via
    #   torch
    #   torchvision
urllib3==1.26.12
    # via requests
